requirements.txt
jmespath==0.10.0
lxml==4.7.1
python-docx==0.8.11
PyYAML==6.0
